requirements.txt
# Core Dependencies
torch>=1.12.0
torchvision>=0.13.0
torch-geometric>=2.1.0

# Computer Vision
opencv-python>=4.6.0
albumentations>=1.3.0
ultralytics>=8.0.0
Pillow>=9.0.0

# Natural Language Processing  
transformers>=4.21.0
tokenizers>=0.13.0

# Scientific Computing
numpy>=1.21.0
scikit-learn>=1.1.0
scipy>=1.9.0

# Data Handling
pandas>=1.4.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Progress Bars and Logging
tqdm>=4.64.0

# Configuration and Utilities
PyYAML>=6.0
argparse
json5

# Optional: Jupyter support
jupyter>=1.0.0
ipywidgets>=7.6.0

# Optional: Tensorboard for logging
tensorboard>=2.8.0

# Optional: Weights & Biases for experiment tracking
# wandb>=0.13.0

# Development Dependencies (optional)
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=0.982

# Note: For GPU support, install appropriate CUDA versions:
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
